third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/worker-interception-redirect.https-expected.txt

This is a testharness.js-based test.
PASS initialize global state
FAIL request to sw1 scope gets network redirect to sw2 scope assert_equals: expected "the shared worker script was served from network" but got "sw2 saw the request for the worker script"
FAIL request to sw1 scope gets network redirect to out-of-scope assert_equals: expected "sw1 saw importScripts from the worker" but got "importScripts: served from network"
FAIL request to sw1 scope gets service-worker redirect to sw2 scope assert_equals: location.href expected "https://web-platform.test:8444/service-workers/service-worker/resources/worker_interception_redirect_webworker.py?greeting=sw2%20saw%20the%20request%20for%20the%20worker%20script" but got "https://web-platform.test:8444/service-workers/service-worker/resources/scope2/worker_interception_redirect_webworker.py"
FAIL request to sw1 scope gets service-worker redirect to out-of-scope assert_equals: expected "sw1 saw importScripts from the worker" but got "importScripts: served from network"
PASS cleanup global state
Harness: the test ran to completion.
